Reference to Manage announcement type APIs
Earlier Object APIs were used to manage Announcement types. Now we have a more targeted set of APIs to do the same. Updated reference to the same. Also removed the clarifications related to the use of object APIs in this context as that is not required with the new manage APIs. Relevant and sufficient information is added to the respective API document.